feat: WinningRace 클래스 생성
1. 결과 비교 책임 부여
2. stream Api filter 사용.
